Jour-2/Apres-midi/VisionAI/vision-app/requirements.txt
Flask==2.3.3
google-cloud-vision==3.4.5
google-cloud-storage==2.10.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0